// src/app/types/generated/tgm.Application.TgAccounts.DTOs.ts
/* THIS (.ts) FILE IS GENERATED BY Tapper */
/* eslint-disable */
/* tslint:disable */
import type { TgServiceState } from './tgm.Domain.TgAccounts.Enums';

/** Transpiled from tgm.Application.TgAccounts.DTOs.AccountDTO */
export type AccountDTO = {
    /** Transpiled from System.Guid */
    id: string;
    /** Transpiled from string */
    firstName: string;
    /** Transpiled from string */
    phoneNumber: string;
    /** Transpiled from tgm.Domain.TgAccounts.Enums.TgServiceState */
    serviceState: TgServiceState;
}

/** Transpiled from tgm.Application.TgAccounts.DTOs.ChatDTO */
export type ChatDTO = {
    /** Transpiled from long */
    id: number;
    /** Transpiled from string */
    title: string;
    /** Transpiled from System.Collections.Generic.List<tgm.Application.TgAccounts.DTOs.ChatPositionDTO> */
    positions: ChatPositionDTO[];
}

/** Transpiled from tgm.Application.TgAccounts.DTOs.ChatFolderInfoDTO */
export type ChatFolderInfoDTO = {
    /** Transpiled from int */
    id: number;
    /** Transpiled from string */
    name: string;
}

/** Transpiled from tgm.Application.TgAccounts.DTOs.ChatListDTO */
export type ChatListDTO = {
    /** Transpiled from string */
    type: string;
    /** Transpiled from int */
    chatFolderId?: number;
}

/** Transpiled from tgm.Application.TgAccounts.DTOs.ChatPositionDTO */
export type ChatPositionDTO = {
    /** Transpiled from tgm.Application.TgAccounts.DTOs.ChatListDTO */
    list: ChatListDTO;
    /** Transpiled from long */
    order: number;
}

/** Transpiled from tgm.Application.TgAccounts.DTOs.MessageDTO */
export type MessageDTO = {
    /** Transpiled from long */
    id: number;
    /** Transpiled from string */
    content: string;
    /** Transpiled from int */
    date: number;
    /** Transpiled from bool */
    isOutgoing: boolean;
}